Moved the display_images function
I also found a bug in the naming of the windows of the cv2.namedWindow() function.

There was a mismatch of names when creating and destroying windows. A window named "images: x" was created and a window named "Images: x" was destroyed? Can you find the error? The "i" was not capitalized when creating. Easy peasy lemon squeezy fix